backend/site-settings/requirements.txt
psycopg2-binary>=2.9.0
PyJWT==2.8.0